// src/components/using-img/IconArt/IconArt.types.ts
export type ThemeType = 'light' | 'dark';

export const iconArtArr: IconArtType[] = [
  'apple',
  'bandages',
  'bell',
  'bicycle',
  'boom',
  'broken-heart',
  'calculator',
  'calendar',
  'car',
  'car-2',
  'car-3',
  'car-4',
  'car-5',
  'car-accident-explosion',
  'car-accident-fire',
  'car-accident-flat-tire',
  'car-accident-lightning',
  'car-accident-rock',
  'car-broken-windshield',
  'car-freshener-tree',
  'car-front',
  'car-light-charger',
  'car-on-cliff',
  'cat-face',
  'clock',
  'convertible-car',
  'court-hammer',
  'credit-card',
  'document',
  'document-2',
  'document-clock',
  'dog-face',
  'dog-paw',
  'driver-with-seat-belt',
  'eaten-shoe',
  'eye-open',
  'flat-tire',
  'foot-cast',
  'keys',
  'magnifying-glass',
  'pill',
  'pill+dog-snack',
  'private-truck',
  'road-cone',
  'sale',
  'spray-can',
  'steering-wheel',
  'stepping-foot',
  'suv',
  'syringe',
  'tooth',
  'usb-stick',
  'user-2',
  'vehicle-warning-sign',
  'wrench'
];

export type IconArtType =
  | 'apple'
  | 'bandages'
  | 'bell'
  | 'bicycle'
  | 'boom'
  | 'broken-heart'
  | 'calculator'
  | 'calendar'
  | 'car'
  | 'car-2'
  | 'car-3'
  | 'car-4'
  | 'car-5'
  | 'car-accident-explosion'
  | 'car-accident-fire'
  | 'car-accident-flat-tire'
  | 'car-accident-lightning'
  | 'car-accident-rock'
  | 'car-broken-windshield'
  | 'car-freshener-tree'
  | 'car-front'
  | 'car-light-charger'
  | 'car-on-cliff'
  | 'cat-face'
  | 'clock'
  | 'convertible-car'
  | 'court-hammer'
  | 'credit-card'
  | 'document'
  | 'document-2'
  | 'document-clock'
  | 'dog-face'
  | 'dog-paw'
  | 'driver-with-seat-belt'
  | 'eaten-shoe'
  | 'eye-open'
  | 'flat-tire'
  | 'foot-cast'
  | 'keys'
  | 'magnifying-glass'
  | 'pill'
  | 'pill+dog-snack'
  | 'private-truck'
  | 'road-cone'
  | 'sale'
  | 'spray-can'
  | 'steering-wheel'
  | 'stepping-foot'
  | 'suv'
  | 'syringe'
  | 'tooth'
  | 'usb-stick'
  | 'user-2'
  | 'vehicle-warning-sign'
  | 'wrench';